Questionnaire

+++ frontend/src/Questionnaire.js
@@ -117,7 +117,7 @@ const Questionnaire = () => {
     console.log(corrected_user_responses);
     let corrected_keys = convertQuestionKeys(corrected_user_responses);
     corrected_keys["free_response"] = description;
-    corrected_keys["budget_constraints"] = parseFloat(budget);
+    corrected_keys["budget_constraints"] = parseInt(budget);
     let report_status = await createReportForUser(corrected_keys, file_name);
   };
